backend/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Celery
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# File handling
Pillow>=10.0,<11.0

# API Documentation
drf-spectacular>=0.27,<0.28

# Utilities
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=22.0,<23.0

# Development
ipython>=8.0,<9.0
